requirements.txt
tqdm
pandas
numpy
sklearn
torch==1.9.0
torchvision==0.10.0
Pillow
matplotlib
timm==0.5.4
protobuf==3.20.1
